+++ dts/prettygoat.d.ts
@@ -21,14 +21,43 @@ declare module prettygoat {
     export interface IWhen<T extends Object> {
         $init?:() => T;
         $any?:(s:T, payload:Object, event?:Event) => T;
-        [name:string]:(s:T, payload:Object, event?:Event) => T;
+        [name:string]:(s:T, payload:Object, event?:Event) => T|SpecialState<T>;
+    }
+
+    abstract class SpecialState<T> {
+        state:T;
+    }
+
+    export class SpecialStates {
+        static stopSignalling<T>(state:T):SpecialState<T>;
+        static deleteSplit():SpecialState<void>;
+    }
+
+    class StopSignallingState<T> extends SpecialState<T> {
+        state:T;
+
+        constructor(state:T);
+    }
+
+    class DeleteSplitState extends SpecialState<void> {
+        state:void;
+
+        constructor();
     }
 
     export interface IProjectionRunner<T> extends IDisposable {
         state:T|Dictionary<T>;
+        stats:ProjectionStats;
         run(snapshot?:Snapshot<T|Dictionary<T>>):void;
         stop():void;
+        pause():void;
+        resume():void;
         notifications:Observable<Event>;
+    }
+
+    export default class ProjectionStats {
+        events:number;
+        readModels:number;
     }
 
     export interface IProjectionRunnerFactory {
@@ -51,6 +80,7 @@ declare module prettygoat {
         initialize():Observable<void>;
         getSnapshots():Observable<Dictionary<Snapshot<any>>>;
         saveSnapshot<T>(streamId:string, snapshot:Snapshot<T>):void;
+        deleteSnapshot(streamId:string):void;
     }
 
     export interface IStreamFactory {