+++ scripts/projections/ProjectionStats.ts
@@ -1,0 +1,4 @@
+export default class ProjectionStats {
+    events:number = 0;
+    readModels:number = 0;
+}