+++ test/ProjectionRunnerSpec.ts
@@ -6,7 +6,7 @@ import {IMatcher} from "../scripts/matcher/IMatcher";
 import {IStreamFactory} from "../scripts/streams/IStreamFactory";
 import {MockMatcher} from "./fixtures/MockMatcher";
 import {MockStreamFactory} from "./fixtures/MockStreamFactory";
-import {Observable, Subject, IDisposable} from "rx";
+import {Observable, Subject, IDisposable, Scheduler} from "rx";
 import * as TypeMoq from "typemoq";
 import expect = require("expect.js");
 import * as Rx from "rx";
@@ -15,6 +15,7 @@ import ReadModelFactory from "../scripts/streams/ReadModelFactory";
 import {Event} from "../scripts/streams/Event";
 import {Snapshot} from "../scripts/snapshots/ISnapshotRepository";
 import MockDateRetriever from "./fixtures/MockDateRetriever";
+import ReservedEvents from "../scripts/streams/ReservedEvents";
 
 describe("Given a ProjectionRunner", () => {
     let stream:TypeMoq.Mock<IStreamFactory>;
@@ -39,7 +40,6 @@ describe("Given a ProjectionRunner", () => {
             }, stream.object, matcher.object, readModelFactory.object, new MockStreamFactory(Observable.empty<Event>()),
             new MockDateRetriever(new Date(100000)));
         subscription = subject.notifications().subscribe((state:Event) => notifications.push(state.payload), e => failed = true, () => stopped = true);
-        readModelFactory.setup(r => r.from(TypeMoq.It.isAny())).returns(_ => Rx.Observable.empty<Event>());
     });
 
     afterEach(() => subscription.dispose());
@@ -53,6 +53,7 @@ describe("Given a ProjectionRunner", () => {
                 splitKey: null
             }));
             matcher.setup(m => m.match(SpecialNames.Init)).returns(streamId => () => 42);
+            readModelFactory.setup(r => r.from(TypeMoq.It.isAny())).returns(_ => Rx.Observable.empty<Event>());
         });
 
         context("if a snapshot is present", () => {
@@ -100,6 +101,7 @@ describe("Given a ProjectionRunner", () => {
     context("when receiving an event from a stream", () => {
         beforeEach(() => {
             matcher.setup(m => m.match(SpecialNames.Init)).returns(streamId => () => 42);
+            readModelFactory.setup(r => r.from(TypeMoq.It.isAny())).returns(_ => Rx.Observable.empty<Event>());
         });
 
         it("it should filter out diagnostic events", () => {
@@ -139,6 +141,10 @@ describe("Given a ProjectionRunner", () => {
                     42 + 1 + 2 + 3 + 4,
                     42 + 1 + 2 + 3 + 4 + 5
                 ]);
+            });
+
+            it("should update the events processed counter", () => {
+                expect(subject.stats.events).to.be(5);
             });
 
             it("should publish on the event stream the new aggregate state", () => {
@@ -189,19 +195,29 @@ describe("Given a ProjectionRunner", () => {
                 expect(failed).to.be.ok();
             });
         });
+    });
 
-        context("and it's the read model of the same projection", () => {
-            let readModelSubject = new Subject<any>();
-            beforeEach(() => {
-                matcher.setup(m => m.match("test")).returns(streamId => (s:number, e:any) => s + e);
-                readModelFactory.setup(s => s.from(null)).returns(_ => readModelSubject);
-                stream.setup(s => s.from(null, TypeMoq.It.isAny())).returns(_ => Observable.empty<Event>());
-                subject.run();
-            });
-
+    context("when receiving a readmodel", () => {
+        let readModelSubject = new Subject<any>();
+        beforeEach(() => {
+            matcher.setup(m => m.match(SpecialNames.Init)).returns(streamId => () => 42);
+            readModelFactory.setup(s => s.from(null)).returns(_ => readModelSubject.observeOn(Scheduler.immediate));
+            stream.setup(s => s.from(null, TypeMoq.It.isAny())).returns(_ => Observable.empty<Event>());
+            subject.run();
+        });
+        context("of the same projection", () => {
+            beforeEach(() => matcher.setup(m => m.match("test")).returns(streamId => (s:number, e:any) => s + e));
             it("should filter it", () => {
                 readModelSubject.onNext({type: "test", payload: 1});
                 expect(subject.state).to.be(42);
+            });
+        });
+
+        context("of another projection", () => {
+            beforeEach(() => matcher.setup(m => m.match("test2")).returns(streamId => (s:number, e:any) => s + e));
+            it("should update the readmodels processed counter", () => {
+                readModelSubject.onNext({type: "test2", payload: 1});
+                expect(subject.stats.readModels).to.be(1);
             });
         });
     });
@@ -209,18 +225,19 @@ describe("Given a ProjectionRunner", () => {
     context("when stopping a projection", () => {
         let streamSubject = new Subject<any>();
         beforeEach(() => {
-            let date = new Date();
+            readModelFactory.setup(r => r.from(TypeMoq.It.isAny())).returns(_ => Rx.Observable.empty<Event>());
             matcher.setup(m => m.match(SpecialNames.Init)).returns(streamId => () => 42);
             matcher.setup(m => m.match("increment")).returns(streamId => (s:number, e:any) => s + e);
             stream.setup(s => s.from(null, TypeMoq.It.isAny())).returns(_ => streamSubject);
 
             subject.run();
-            streamSubject.onNext({type: "increment", payload: 1, timestamp: new Date(+date + 1)});
-            streamSubject.onNext({type: "increment", payload: 2, timestamp: new Date(+date + 2)});
-            streamSubject.onNext({type: "increment", payload: 3, timestamp: new Date(+date + 3)});
-            streamSubject.onNext({type: "increment", payload: 4, timestamp: new Date(+date + 4)});
+            streamSubject.onNext({type: ReservedEvents.REALTIME, payload: null, timestamp: null});
+            streamSubject.onNext({type: "increment", payload: 1, timestamp: new Date(501)});
+            streamSubject.onNext({type: "increment", payload: 2, timestamp: new Date(502)});
+            streamSubject.onNext({type: "increment", payload: 3, timestamp: new Date(503)});
+            streamSubject.onNext({type: "increment", payload: 4, timestamp: new Date(504)});
             subject.stop();
-            streamSubject.onNext({type: "increment", payload: 5, timestamp: new Date(+date + 5)});
+            streamSubject.onNext({type: "increment", payload: 5, timestamp: new Date(505)});
         });
         it("should not process any more events", () => {
             expect(notifications).to.eql([
@@ -238,6 +255,44 @@ describe("Given a ProjectionRunner", () => {
         context("and the projection is started again", () => {
             it("should throw an error", () => {
                 expect(() => subject.run()).to.throwError();
+            });
+        });
+    });
+
+    context("when pausing a projection", () => {
+        let streamSubject = new Subject<any>();
+        beforeEach(() => {
+            readModelFactory.setup(r => r.from(TypeMoq.It.isAny())).returns(_ => Rx.Observable.empty<Event>());
+            matcher.setup(m => m.match(SpecialNames.Init)).returns(streamId => () => 42);
+            matcher.setup(m => m.match("increment")).returns(streamId => (s:number, e:any) => s + e);
+            stream.setup(s => s.from(null, TypeMoq.It.isAny())).returns(_ => streamSubject);
+
+            subject.run();
+            streamSubject.onNext({type: ReservedEvents.REALTIME, payload: null, timestamp: null});
+            streamSubject.onNext({type: "increment", payload: 1, timestamp: new Date(501)});
+            streamSubject.onNext({type: "increment", payload: 2, timestamp: new Date(502)});
+            subject.pause();
+            streamSubject.onNext({type: "increment", payload: 3, timestamp: new Date(503)});
+        });
+        it("should not process events anymore", () => {
+            expect(notifications).to.eql([
+                42,
+                42 + 1,
+                42 + 1 + 2
+            ]);
+        });
+
+        context("and the projection is resumed", () => {
+            it("should start from the last state", () => {
+                subject.resume();
+                streamSubject.onNext({type: "increment", payload: 4, timestamp: new Date(504)});
+                expect(notifications).to.eql([
+                    42,
+                    42 + 1,
+                    42 + 1 + 2,
+                    42 + 1 + 2 + 4
+                ]);
+                expect(subject.state).to.be(42 + 1 + 2 + 4);
             });
         });
     });