+++ test/TimePartitionerSpec.ts
@@ -1,5 +1,4 @@
 import expect = require("expect.js");
-import sinon = require("sinon");
 import TimePartitioner from "../scripts/util/TimePartitioner";
 import MockDateRetriever from "./fixtures/MockDateRetriever";